AnnotationRenderService - Now relies on events for rendering and clearing canvas
